+++ plugins/alignments/src/PileupRPC/methods/GetGlobalValueForTag.ts
@@ -1,0 +1,59 @@
+import RpcMethodType from '@jbrowse/core/pluggableElementTypes/RpcMethodType'
+import { getAdapter } from '@jbrowse/core/data_adapters/dataAdapterCache'
+import { renameRegionsIfNeeded, Region } from '@jbrowse/core/util'
+import { RenderArgs } from '@jbrowse/core/rpc/coreRpcMethods'
+import { RemoteAbortSignal } from '@jbrowse/core/rpc/remoteAbortSignals'
+import { BaseFeatureDataAdapter } from '@jbrowse/core/data_adapters/BaseAdapter'
+import { toArray } from 'rxjs/operators'
+
+export default class PileupGetGlobalValueForTag extends RpcMethodType {
+  name = 'PileupGetGlobalValueForTag'
+
+  async serializeArguments(
+    args: RenderArgs & {
+      signal?: AbortSignal
+      statusCallback?: (arg: string) => void
+    },
+    rpcDriver: string,
+  ) {
+    const { rootModel } = this.pluginManager
+    const assemblyManager = rootModel?.session?.assemblyManager
+    if (!assemblyManager) {
+      throw new Error('no assembly manager available')
+    }
+
+    const renamedArgs = await renameRegionsIfNeeded(assemblyManager, args)
+
+    return super.serializeArguments(renamedArgs, rpcDriver)
+  }
+
+  async execute(
+    args: {
+      adapterConfig: {}
+      signal?: RemoteAbortSignal
+      headers?: Record<string, string>
+      regions: Region[]
+      sessionId: string
+      tag: string
+    },
+    rpcDriver: string,
+  ) {
+    const { adapterConfig, sessionId, regions, tag } =
+      await this.deserializeArguments(args, rpcDriver)
+
+    const dataAdapter = (
+      await getAdapter(this.pluginManager, sessionId, adapterConfig)
+    ).dataAdapter as BaseFeatureDataAdapter
+
+    const features = dataAdapter.getFeaturesInMultipleRegions(regions)
+    const featuresArray = await features.pipe(toArray()).toPromise()
+    return [
+      ...new Set(
+        featuresArray
+          .map(feature => getTag(feature, tag))
+          .filter(f => f !== undefined)
+          .map(f => `${f}`),
+      ),
+    ]
+  }
+}
